Fix first run of check-zip command incorrectly returning no ZIP libraries are available. Catch and display various errors that can occur while checking for availability of ZIP libraries.

+++ src/Console/Commands/CheckZipLibraries.php
@@ -41,45 +41,69 @@ class CheckZipLibraries extends Command
     {
 		$this->info("Checking Zip libraries availability...");
 
+		if (!$this->validateZipPassword()) {
+			$this->info("ZIP password hasn't been set. Please check the centinelApi.php config file, or run 'php artisan centinel-api:setup' console command.");
+
+			return;
+		}
+
 		$filePath = LogFile::getOrCreateFirstExistingLogFilePath();
 		$zipPath = storage_path('logs/laravel.zip');
 
-		Zipper::createNativeZip($filePath, $zipPath);
+		$errorMessage = Zipper::createNativeZip($filePath, $zipPath);
 
 		if (file_exists($zipPath)) {
 			unlink($zipPath);
 
-			$this->info("You're using PHP version " . PHP_VERSION . " so native Zip encryption is available! Your database dumps will be zipped and encrypted with AES-256.");
+			if (!$errorMessage) {
+				$this->info("You're using PHP version " . PHP_VERSION . " so native Zip encryption is available! Your database dumps will be zipped and encrypted with AES-256.");
 
-			return;
+				return;
+			}
 		}
 
-		$this->info("Native Zip encryption is not available.");
+		$this->info($errorMessage ?? "Native Zip encryption is not available.");
 
-		Zipper::create7zip($filePath, $zipPath);
+		$errorMessage = Zipper::create7zip($filePath, $zipPath);
 
 		if (file_exists($zipPath)) {
 			unlink($zipPath);
 
-			$this->info("7-zip is available! Your database dumps will be zipped using 7-Zip and encrypted with AES-256.");
+			if (!$errorMessage) {
+				$this->info("7-zip is available! Your database dumps will be zipped using 7-Zip and encrypted with AES-256.");
 
-			return;
+				return;
+			}
 		}
 
-		$this->info("7-zip is not available.");
+		$this->info($errorMessage ?? "7-zip is not available.");
 
-		Zipper::createRegularZip($filePath, $zipPath);
+		$errorMessage = Zipper::createRegularZip($filePath, $zipPath);
 
 		if (file_exists($zipPath)) {
 			unlink($zipPath);
 
-			$this->info("Zip is available! Your database dumps will be zipped using Zip library and protected using password from the Centinel API config file.");
-			$this->info("It is your responsibility to read up on Zip password protection and decide if this level of security is satisfactory.");
+			if (!$errorMessage) {
+				$this->info("Zip is available! Your database dumps will be zipped using Zip library and protected using password from the Centinel API config file.");
+				$this->info("It is your responsibility to read up on Zip password protection and decide if this level of security is satisfactory.");
 
-			return;
+				return;
+			}
 		}
 
-		$this->info("Zip library is not available.");
-		$this->info("Your database dumps will be sent to Centinel without being zipped and password protected beforehand.");
+		$this->info($errorMessage ?? "Zip library is not available.");
+
+		$this->info(
+			$errorMessage ?
+			"Please fix the errors listed above and run the 'php artisan centinel-api:check-zip' command again." :
+			"Your database dumps will be sent to Centinel without being zipped and password protected beforehand."
+		);
+
+		return;
+    }
+
+	protected function validateZipPassword()
+	{
+		return config('centinelApi.zipPassword') ? true : false;
     }
 }
